Updates for phase 3
Adding configuration directory as input parameter to finalize().

+++ 1N_2018/validation/scripts/run_testdriver.sh
@@ -53,7 +53,9 @@ chmod -R 550 $configDir
 outputDir=validation
 outputStem=validation
 enrollDir=$outputDir/enroll
-chmod -R 777 $outputDir; rm -rf $outputDir 
+if [ -d $outputDir ]; then
+	chmod -R 777 $outputDir; rm -rf $outputDir 
+fi
 mkdir -p $enrollDir
 
 root=$(pwd)
